refactored configuration interfaces

+++ sources/src/main/java/eu/unifiedviews/plugins/transformer/fusiontool/config/xml/FileOutputsXml.java
@@ -3,11 +3,11 @@
  */
 package eu.unifiedviews.plugins.transformer.fusiontool.config.xml;
 
-import java.util.List;
-
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.ElementList;
 import org.simpleframework.xml.Root;
+
+import java.util.List;
 
 // CHECKSTYLE:OFF
 
@@ -17,12 +17,12 @@ import org.simpleframework.xml.Root;
 @Root(name = "FileOutputsXml")
 public class FileOutputsXml {
     @Attribute(required = false)
-    private Integer maxResolvedQuads;
+    private Long maxResolvedQuads;
     
     @ElementList(name = "FileOutput", required = false, inline = true, empty = false)
     private List<FileOutputXml> fileOutputs;
 
-    public Integer getMaxResolvedQuads() {
+    public Long getMaxResolvedQuads() {
         return maxResolvedQuads;
     }
